# backend/booking-service/docs/swagger.yaml
basePath: /
definitions:
...
  internal_handler.createCinemaRequest:
    properties:
      address:
        type: string
      city:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      phone:
        type: string
      rating:
        type: number
    type: object
  internal_handler.createHallRequest:
    properties:
      cinema_id:
        example: 1
        type: integer
      name:
        example: VIP Hall 1
        type: string
      rows:
        example: 10
        type: integer
      seats_per_row:
        example: 10
        type: integer
    type: object
  internal_handler.createMovieRequest:
    properties:
      banner_url:
        example: https://example.com/banner.jpg
        type: string
      description:
        example: Epic sci-fi film
        type: string
      duration_minutes:
        example: 166
        type: integer
      poster_url:
        example: https://example.com/poster.jpg
        type: string
      rating:
        example: 8.7
        type: number
      release_date:
        example: "2024-03-01T00:00:00Z"
        type: string
      title:
        example: 'Dune: Part Two'
        type: string
      trailer_url:
        example: https://example.com/trailer.mp4
        type: string
    type: object
...
  internal_handler.sessionResponse:
    properties:
      base_price_cents:
        type: integer
      hall_id:
        type: integer
      id:
        type: integer
      movie_id:
        type: integer
      start_time:
        type: string
    type: object
  internal_handler.updateCinemaRequest:
    properties:
      address:
        type: string
      city:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      phone:
        type: string
      rating:
        type: number
    type: object
  internal_handler.updateMovieRequest:
    properties:
      banner_url:
        example: https://example.com/banner.jpg
        type: string
      description:
        example: Epic sci-fi film
        type: string
      duration_minutes:
        example: 166
        type: integer
      poster_url:
        example: https://example.com/poster.jpg
        type: string
      rating:
        example: 8.7
        type: number
      release_date:
        example: "2024-03-01T00:00:00Z"
        type: string
      title:
        example: 'Dune: Part Two'
        type: string
      trailer_url:
        example: https://example.com/trailer.mp4
        type: string
    type: object
...
  model.Cinema:
    properties:
      address:
        type: string
      city:
        type: string
      created_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      integration_level:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      phone_number:
        type: string
      rating:
        type: number
    type: object
  model.Hall:
    properties:
      cinema_id:
        type: integer
      created_at:
        type: string
      hall_type:
        type: string
      id:
        type: integer
      name:
        type: string
      rows_count:
        type: integer
      seats_per_row:
        type: integer
      total_seats:
        type: integer
    type: object
  model.Movie:
    properties:
      banner_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      duration_minutes:
        type: integer
      id:
        type: integer
      poster_url:
        type: string
      rating:
        type: number
      release_date:
        type: string
      title:
        type: string
      trailer_url:
        type: string
    type: object
...
info:
  contact: {}
  description: API for movie sessions and bookings
  title: Booking Service API
  version: "1.0"
paths:
  /admin/cinemas:
    get:
      description: Returns all cinemas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Cinema'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get cinemas
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Creates a new cinema
      parameters:
      - description: Cinema payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createCinemaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Cinema'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create cinema
      tags:
      - admin
  /admin/cinemas/{cinemaId}/halls:
    get:
      description: Returns all halls for cinema
      parameters:
      - description: Cinema ID
        in: path
        name: cinemaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Hall'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get halls by cinema
      tags:
      - admin
  /admin/cinemas/{id}:
    delete:
      description: Deletes cinema by ID
      parameters:
      - description: Cinema ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete cinema
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Updates cinema by ID
      parameters:
      - description: Cinema ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cinema payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateCinemaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Cinema'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update cinema
      tags:
      - admin
  /admin/halls:
    post:
      consumes:
      - application/json
      description: Creates a new hall
      parameters:
      - description: Hall payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createHallRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Hall'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create hall
      tags:
      - admin
  /admin/movies:
    get:
      description: Returns all movies for admin management
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Movie'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get admin movies
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Creates a new movie
      parameters:
      - description: Movie payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createMovieRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Movie'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create movie
      tags:
      - admin
  /admin/movies/{id}:
    delete:
      description: Deletes movie by ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete movie
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Updates movie by ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      - description: Movie payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateMovieRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Movie'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update movie
      tags:
      - admin
  /admin/sessions:
    get:
      description: Returns all sessions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handler.sessionResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get sessions
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Creates a new session
      parameters:
      - description: Session payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.sessionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create session
      tags:
      - admin
...
swagger: "2.0"